missions, glide, some other stuff

+++ app/src/main/java/com/grappim/spacexapp/model/rocket/Height.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Height(
   @SerializedName("feet")
-  val feet: Int?,
+  val feet: Double?,
   @SerializedName("meters")
   val meters: Double?
 )
